requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.31.0
langchain>=0.1.0
langchain-google-genai>=0.0.9
langgraph>=0.0.22
python-dotenv>=1.0.1
pydantic>=2.6.0
requests>=2.31.0
email-validator>=2.1.0
langsmith>=0.1.0
langchain-groq>=0.1.0
PyPDF2>=3.0.1
